+++ src/models/StyledComponent.js
@@ -112,6 +112,7 @@ export default (ComponentStyle: Function) => {
         propsForElement.className = [className, componentId, classes, generatedClassName].filter(x => x).join(' ')
         if (innerRef) {
           propsForElement.ref = innerRef
+          delete propsForElement.innerRef
         }
 
         return createElement(target, propsForElement, children)