+++ CHANGELOG.md
@@ -12,12 +12,20 @@ All notable changes to this project will be documented in this file. If a contri
 - Added chained `classes` and `displayName` constructor functions.
 - Added the ability to override one component's styles from another.
 - Injecting an empty class for each instance of a component.
+- Added [`withTheme`](docs/api.md#withtheme) higher order component; thanks [@brunolemos](https://twitter.com/brunolemos). (see [#312] (https://github.com/styled-components/styled-components/pull/312))
+- Fixed prop changes not updating style on react native; thanks [@brunolemos](https://twitter.com/brunolemos). (see [#311](https://github.com/styled-components/styled-components/pull/311))
 
 ### Changed
 
-- Fixed theme changes using `<ThemeProvider />` where classNames didn't changed at the component, thanks [@k15a](https://github.com/k15a)
+- Do not pass innerRef to the component, thanks [@mkhazov](https://github.com/mkhazov). (see [#310](https://github.com/styled-components/styled-components/pull/310))
+
+## [v1.1.3]
+
+### Changed
+
+- Fixed theme changes in `ThemeProvider`s not re-rendering correctly, thanks [@k15a](https://github.com/k15a). (see [#264](https://github.com/styled-components/styled-components/pull/264))
+- Fixed overriding theme through props, thanks [@k15a](https://github.com/k15a). (see [#295](https://github.com/styled-components/styled-components/pull/295))
 - Removed `lodash` dependency in favor of small utility packages to knock down bundle size by ~0.5kB
-- Fixed overriding theme through props, thanks [@k15a](https://github.com/k15a)
 
 ## [v1.1.2]
 
@@ -126,7 +134,8 @@ All notable changes to this project will be documented in this file. If a contri
 
 - Fixed compatibility with other react-broadcast-based systems (like `react-router` v4)
 
-[Unreleased]: https://github.com/styled-components/styled-components/compare/v1.1.2...master
+[Unreleased]: https://github.com/styled-components/styled-components/compare/v1.1.3...master
+[v1.1.3]: https://github.com/styled-components/styled-components/compare/v1.1.2...v1.1.3
 [v1.1.2]: https://github.com/styled-components/styled-components/compare/v1.1.1...v1.1.2
 [v1.1.1]: https://github.com/styled-components/styled-components/compare/v1.1.0...v1.1.1
 [v1.1.0]: https://github.com/styled-components/styled-components/compare/v1.0.11...v1.1.0